Move IdLookup back to common as it's not useful in the API
